add swagger + add exceptions filter

+++ src/infrastructure/entities/truck.entity.ts
@@ -21,6 +21,6 @@ export class Truck {
   })
   location: LocationEnum;
 
-  @Column({type: "decimal"})
+  @Column({type: "float"})
   score: number;
 }
